rename files, add translation form_choice
rename form1 to form_soft_type, and add translation to forms_choice +
rename labels of form_soft_type
