+++ softwareInc_mod_exe v0.4_3_2/softwareInc_mod_exe/bin/Debug/softwareInc_mod_exe.exe.manifest
@@ -43,14 +43,14 @@
     </dependentAssembly>
   </dependency>
   <dependency>
-    <dependentAssembly dependencyType="install" allowDelayedBinding="true" codebase="softwareInc_mod_exe.exe" size="4535808">
+    <dependentAssembly dependencyType="install" allowDelayedBinding="true" codebase="softwareInc_mod_exe.exe" size="4536832">
       <assemblyIdentity name="softwareInc_mod_exe" version="1.0.0.0" language="neutral" processorArchitecture="msil" />
       <hash>
         <dsig:Transforms>
           <dsig:Transform Algorithm="urn:schemas-microsoft-com:HashTransforms.Identity" />
         </dsig:Transforms>
         <dsig:DigestMethod Algorithm="http://www.w3.org/2000/09/xmldsig#sha256" />
-        <dsig:DigestValue>o/9CSR7g8VvkhJhFdH4b5pXDKwhvnkxb8z9vWSlYhoE=</dsig:DigestValue>
+        <dsig:DigestValue>+cDZEfKsqUjfiEXhXhOwEG1Ib69WqAl84IDy2++t83A=</dsig:DigestValue>
       </hash>
     </dependentAssembly>
   </dependency>
