+++ softwareInc_mod_exe v0.4_3_2/softwareInc_mod_exe/bin/Debug/softwareInc_mod_exe.vshost.application
@@ -14,7 +14,7 @@
           <dsig:Transform Algorithm="urn:schemas-microsoft-com:HashTransforms.Identity" />
         </dsig:Transforms>
         <dsig:DigestMethod Algorithm="http://www.w3.org/2000/09/xmldsig#sha256" />
-        <dsig:DigestValue>xlQowJ6lzsKcdWrwGeetjUDdlX9zhu+LNZLoMX5/r7U=</dsig:DigestValue>
+        <dsig:DigestValue>ARdXw9/S8vfD1vYwB+Z3sdLKjvdKV/hiDyPv/oJ7/wQ=</dsig:DigestValue>
       </hash>
     </dependentAssembly>
   </dependency>
